% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{cond_reg}
\alias{cond_reg}
\title{The conditional correlation approach to detecting the complementarity}
\usage{
cond_reg(dat)
}
\arguments{
\item{dat}{A simulated dataset with 5 columns}
}
\value{
A list with the sur object and the name of the method
}
\description{
The conditional correlation approach to detecting the complementarity
}
\seealso{
match_reg run_regression format_reg

Other tests: \code{\link{interaction_reg}},
  \code{\link{match_reg}}, \code{\link{run_regression}},
  \code{\link{sur_reg}}
}

