% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_sample}
\alias{create_sample}
\title{Create a sample of surviving observations for simulation}
\usage{
create_sample(obs = 200, rate = 0.5, sd = 1, b = c(0, 1, 1, 1),
  d = c(0.5, 0.5))
}
\arguments{
\item{obs}{The number of observations in the dataset}

\item{rate}{The survival rate. This should be a number between 0 and 1. The
closer to 0 the stronger the pressure to be optimal.}

\item{sd}{The standard deviation of noise term of the performance. A higher
\code{sd} means that the decisions of interest are less important for
performance.}

\item{b}{The contribution to performance of the two decisions of interest. A
four element vector for the intercept, \code{x1, x2, x1:x2}}

\item{d}{The quadratic effect of the two decisions that captures their the
diminishing returns.}
}
\value{
A dataset with simulated \code{y} and \code{x1}, \code{x2} for
  surviving observations. The data has dimensions = \code{c(obs, 7)}.
}
\description{
Create a sample of surviving observations for simulation
}

