% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{format_reg}
\alias{format_reg}
\title{Transform the regression and method list into a vector.}
\usage{
format_reg(list_reg)
}
\arguments{
\item{list_reg}{A list with two elements. The regression object and a string
with the name of the methodology.}
}
\value{
A row vector with type of analysis, coefficient, t - statistic, and
p - value of the complementarity test.
}
\description{
Transform the regression and method list into a vector.
}

