% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{match_reg}
\alias{match_reg}
\title{Perform the traditional single equation regression approach under the
assumption that \code{x1, x2} are optimally chosen.}
\usage{
match_reg(dat)
}
\arguments{
\item{dat}{A simulated dataset with 7 columns.}
}
\value{
A list with the regression object and the name of the method
}
\description{
Perform the traditional single equation regression approach under the
assumption that \code{x1, x2} are optimally chosen.
}
\seealso{
Other tests: \code{\link{interaction_reg}},
  \code{\link{run_regression}}
}

