% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_regressions.R
\name{run_regression}
\alias{run_regression}
\title{A general function to run all possible regression based approaches.}
\usage{
run_regression(dat, family = "match", method = "traditional")
}
\arguments{
\item{dat}{A simulated dataset with 7 columns.}

\item{family}{"match" for the approach where the optimal choice of the
\code{x}'s is chosen and "interaction" where the \code{x}'s are assumed
to be chosen optimally.}

\item{method}{"traditional" for the typical model with interaction,
'augmented' for the model with interaction and quadratic effects, and
'sur' for the seemingly unrelated regression method.
\code{method} is ignored when \code{family = "match"} or
\code{family = "sur"}.}
}
\value{
A list with the regression object and the name of the method
}
\description{
A general function to run all possible regression based approaches.
}
\seealso{
match_reg interaction_regression format_reg

Other tests: \code{\link{interaction_reg}},
  \code{\link{match_reg}}, \code{\link{sur_reg}}
}

